requirements.txt
tensorflow>=2.8.0
numpy
pandas
folium
requests
python-dotenv
